PB-903 Add function to get class field from language suffix
Adding this function to a child class of django.db.models.Model did not work because models need to be registered. It is an ugly way but I don't see how Django models can be more structured to avoid such flat lists of field names. At least this is going to work for any model class, not just Provider (e.g. distributions.models.Attribution).