PB-903 Drop the explicit schema invocations
According to the docs, it isn't necessary to invoke the Schemas by
ourselves. The endpoint functions return the data as lists, querysets or
dicts and the Framework handles the conversion to the schema. I don't
see an immediate drawback to our version, but I'm also not so
comfortable to deviate from the standard way proposed by the docs.

From the [docs](https://django-ninja.dev/guides/response/), NinjaAPI
seems to do quite some tasks with the returned data itself and in my
opinion we shouldn't be doing this ourselves unless we have a good
reason to.
Also, there might be performance optimizations being done by NinjaAPI
(consider especially when returning entire querysets)