PB-1045 Create dedicated NinjaAPI object for "/checker"
This fixes unit test CheckerUrlTestCase.test_checker_url that reported error "ninja.errors.ConfigError: Router@'' has already been attached to API NinjaAPI:1.0.0" when run in parallel with other tests.

A reason for the error might have been that this test introduced a mix of the django test client and the ninja test client. The ninja test client test skips the middleware/url-resolver layer while the django test client does not. So somehow Ninja's router registration was done twice.

Just using the ninja test client wasn't enough then because it didn't know the "/checker" endpoint, which was only defined in the "urls" module. So the "/checker" endpoint is now also handled by its own NinjaAPI object.

In the course of this, the definition of the "/checker" endpoint was moved back to the "api" module where API endpoints are usually defined.