B-1805: Fixed logout for un-authorized user
Some user might have successfully authenticate but are not authorized to access
the admin interface. Those needs to be able to logout. That's why the `@staff_member_required`
is removed from the `custom_admin_logout` endpoint and that a logout button
is added to the login page.

Also don't breaks the oauth2-proxy middleware when some infos are missing.